backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9  # Replace mysqlclient
python-socketio==5.11.0
Werkzeug==2.3.7
gunicorn==22.0.0
eventlet==0.36.1